Data.SciRatio

Description

The SciRatio type

data SciRatio a b

Represents a fractional number stored in scientific notation: a product of a fractional significand and an integral power of 10.

The significand has type a and should be both Fractional and Real. Although this could be a floating-point type, it is not recommended as floating-points use inexact arithmetic and strange bugs will occur as a result.

The exponent has type b and should be Integral.

SciRatio behaves in the same way as an ordinary Ratio and supports the same operations. The main property is that it is moxre efficient than Ratio when the exponent is large:

>>> read "(5 % 1) .^ 99999999" :: SciRational   -- works fine
>>> 5e99999999                 :: Rational      -- takes forever

Specialized functions are provided in cases where they can be implemented more efficiently than the default implementation.

The number is always stored in a unique, canonical form: the significand shall never contain factors of 2 and 5 simultaneously, and their multiplicities shall always be nonnegative. (Note that here we treat the significand as a rational number factorized into a product of prime numbers with integral exponents.)

Note: If inputs differ greatly in magnitude, (+) and (-) can be quite slow: complexity is linear with the absolute difference of the exponents. Furthermore, the complexity of toRational is linear with the magnitude of the exponent. These also apply to any functions that indirectly use these operations (including all operations in RealFrac and Enum).

type SciRational = SciRatio Rational Integer

A specialization of SciRatio.

(.^) infixl 7

Arguments

:: (Fractional a, Real a, Integral b)

=> a
significand

-> b
exponent

-> SciRatio a b

Construct a number such that significand .^ exponent == significand * 10 ^^ exponent.

fracSignificand :: SciRatio a b -> a

Extract the fractional significand.

base10Exponent :: SciRatio a b -> b

Extract the base-10 exponent.

Specialized functions

(^!) :: (Real a, Integral b, Integral c) => SciRatio a b -> c -> SciRatio a b infixr 8

Specialized, more efficient version of (^).

(^^!) :: (Fractional a, Real a, Integral b, Integral c) => SciRatio a b -> c -> SciRatio a b infixr 8

Specialized, more efficient version of (^^).

fromSciRatio :: (Real a, Integral b, Fractional c) => SciRatio a b -> c

Convert into a Fractional number.

This is similar to realToFrac but much more efficient for large exponents.

Miscellaneous utilities

intLog

Arguments

:: (Integral a, Integral b)

=> a
base

-> a
input integer

-> (a, b)
significand and exponent

Extract the largest power of the given base that divides the input integer. Returns the significand and exponent, satisfying:

input_integer = significand * base ^ exponent

If the input integer is zero, then zeros are returned.

If the input integer is negative, the significand is negative.

This is similar to computing the floored logarithm and its associated remainder.