Data.SciRatio.Utils

Description

Pretty-printers

showNumber :: Real a => a -> String

Show a number (see showsNumber).

Note: for SciRational, consider using the more efficient, specialized function showSciRational instead.

showSciRational :: SciRational -> String

Show a number (see showsNumber).

showsNumber :: Real a => a -> ShowS

Show a rational number in scientific notation:

[-+]?
( [0-9]+ [.]? [0-9]* | [.] [0-9]+ )
( [e] [-+]? [0-9]+ )?
( [/] [0-9]+ )?

Note: for SciRational, consider using the more efficient, specialized function showsSciRational instead.

showsSciRational :: SciRational -> ShowS

Show a number (see showNumber).

Parsers

readNumber :: Fractional a => String -> Maybe a

Read a number (see readNumberP).

readSciRational :: String -> Maybe SciRational

Read a number (see readNumberP).

readSign :: Num a => Char -> Maybe a

Interpret a sign:

sign = [-+]

Note: U+2212 (MINUS SIGN) is also accepted.

readNumberP :: Fractional a => ReadP a

Read a rational number in scientific notation:

number = [0] [bB] [0-1]+
       | [0] [oO] [0-7]+
       | [0] [xX] [0-9a-fA-F]+
       | scientific ( fraction_slash scientific )?

readSciRationalP :: ReadP SciRational

Read a number (see readNumberP).

readScientificP :: Fractional a => ReadP a

Read a decimal fraction in scientific notation:

scientific = decimal (decimal_exponent_symbol sign? dec)?

readDecimalP :: Fractional a => ReadP a

Read a decimal fraction:

decimal = sign? ( [0-9]+ [.]? [0-9]*
                | [.] [0-9]+ )

readIntegerP :: Num a => ReadP a

Read a signed integer in base 10.

integer = sign? [0-9]+

readUnsignedP :: Num a => ReadP a

Read a unsigned number in either binary (0b), octal (0o), decimal, or hexadecimal (0x) format:

unsigned = [0] [bB] [0-1]+
         | [0] [oO] [0-7]+
         | [0] [xX] [0-9a-fA-F]+
         | dec

readHexP :: Num a => ReadP a

Read an unsigned integer in hexadecimal notation:

hex = [0-9A-Fa-f]+

readDecP :: Num a => ReadP a

Read an unsigned integer in base 10.

dec = [0-9]+

Note: Although similar functions exist in Lex, the versions here do not require Eq.

readOctP :: Num a => ReadP a

Read an unsigned integer in octal notation:

oct = [0-7]+

readBinP :: Num a => ReadP a

Read an unsigned integer in binary notation:

bin = [01]+

Character predicates

isBinDigit :: Char -> Bool

Whether the character is a binary digit:

bin_digit = [01]

isDecimalExponentSymbol :: Char -> Bool

Whether the character can be used as an exponent symbol in scientific notation:

decimal_exponent_symbol = [eE]

Note: U+23E8 (DECIMAL EXPONENT SYMBOL) is also accepted.

isFractionSlash :: Char -> Bool

Whether the character is a fraction slash:

fraction_slash = [/]

Note: U+002F (SOLIDUS), U+2044 (FRACTION SLASH), and U+2215 (DIVISION SLASH) are also accepted.